refactor(WinningResult) : WinningResult 관련 책임 이동 LadderGame -> WinningResult (LadderGame 삭제)
